password modal connect to redux
